made changes to fix modal not presenting properly
Per Kevin Powell, having the modal section inside of a div containing a backdrop-filter was preventing the modal from presenting properly as fixed in the viewport.
To resolve this issue the image using backdrop-filter: blur(3px) was taken out and manually blurred using Corel Paintshop Pro and the backdrop-filter:blur(3px) was removed from the site.
Verified the modal now presents correctly.